% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bedtoolsGenomeCov.R
\name{bedtoolsGenomeCov}
\alias{bedtoolsGenomeCov}
\title{Calculate coverage of a set of bed files at each base of the genome.}
\usage{
bedtoolsGenomeCov(bedFiles, genomefile, covmode = "-bg", scalar = "rpm",
  bam = FALSE, blocks = FALSE, threads = getOption("threads", 1L))
}
\arguments{
\item{bedFiles}{A character vector of paths to bed files.}

\item{scalar}{A number to multiply the read counts by. If "rpm", the read counts will be multiplied by a 1000000/(number of reads) to yield reads-per-million (rpm).}

\item{threads}{A positive integer specifying how many samples to process simultaneously.}

\item{windowfile}{A string specifying the path to the bed file within which to calculate coverage of beds.}

\item{windowsize}{Positive integer indicating size of windows used to resize overlapping windows to comply with bedGraph format restriction against overlapping intervals. This only needs to be specified if windows are overlapping and a bedGraph-compliant file is desired.}

\item{stepsize}{Positive integer indicating size of the steps used to resize overlapping windows to comply with bedGraph format restriction against overlapping intervals. This only needs to be specified if windows are overlapping and a bedGraph-compliant file is desired.}
}
\description{
\code{bedtoolsGenomeCov} is a wrapper for \code{bedtools genomecov -bg}.
}

