% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bedtoolsMap.R
\name{bedtoolsMap}
\alias{bedtoolsMap}
\title{Perform calculations on bedGraph scores over the inverals in a bed file.}
\usage{
bedtoolsMap(bgFiles, windows, operation = "mean", outnames = NULL,
  windowsize = 25, stepsize = windowsize, filler = 0, printzero = TRUE,
  threads = getOption("threads", 1L))
}
\arguments{
\item{windows}{A string specifying the path to the bed file within which to perform calculations on bedGraph files.}

\item{operation}{to perform on bedGraph intervals overlapping windows. Can be one of: sum,count,count_distinct,min,max,absmin,absmax,mean,median,antimode,collapse,distinct,concat. Default is mean. See bedtools docs for more info.}

\item{outnames}{A character vector of output file names. If NULL, output file names are automatically generated using the syntax bedGraphFileName_operation_windowFileName.bg}

\item{windowsize}{Positive integer indicating size of windows. This is used to resize overlapping windows to comply with bedGraph format restriction against overlapping intervals. This only needs to be specified if windows are overlapping and a bedGraph-compliant file is desired.}

\item{stepsize}{Positive integer indicating size of the steps. This is used to resize overlapping windows to comply with bedGraph format restriction against overlapping intervals. This only needs to be specified if windows are overlapping and a bedGraph-compliant file is desired.}

\item{filler}{Value to assign intervals with no overlapping bedGraph intervals.}

\item{printzero}{Boolean indicating of regions with a value of 0 should be printed. Default is TRUE.}

\item{threads}{A positive integer specifying how many samples to process simultaneously.}

\item{bgFiless}{A character vector of paths to bedGraph files.}
}
\description{
\code{bedtoolsMap} is a wrapper for \code{bedtools map} that "allows one to map overlapping features in a B file onto features in an A file and apply statistics and/or summary operations on those features."
}

