% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bedtoolsSubtract.R
\name{bedtoolsSubtract}
\alias{bedtoolsSubtract}
\title{Remove regions from a bed file overlapping another bed file.}
\usage{
bedtoolsSubtract(a, b, extraargs = "", header = TRUE, count = FALSE,
  threads = getOption("threads", 1L))
}
\arguments{
\item{a}{A character vector of paths to bed files.}

\item{b}{A path to the bed file used to remove intervals in a.}

\item{extraargs}{A string passed as arguments to bedtools subtract}

\item{count}{If TRUE, only returns the number of intervals from the operation.}

\item{threads}{A positive integer specifying how many samples to process simultaneously.}
}
\description{
\code{bedtoolsSubtract} is a wrapper for \code{bedtools subtract}.
}

