% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gyro.R
\encoding{UTF-8}
\name{gyrocentroid}
\alias{gyrocentroid}
\title{Gyrocentroid}
\usage{
gyrocentroid(A, B, C, s = 1, model = "U")
}
\arguments{
\item{A, B, C}{three distinct points}

\item{s}{positive number, the radius of the Poincaré ball if
\code{model="M"}, otherwise, if \code{model="U"}, this number
defines the hyperbolic curvature (the smaller, the more curved)}

\item{model}{the hyperbolic model, either \code{"M"} (Möbius model, i.e.
Poincaré model) or \code{"U"} (Ungar model, i.e. hyperboloid model)}
}
\value{
A point, the gyrocentroid of the triangle \code{ABC}.
}
\description{
Gyrocenroid of a triangle.
}
