% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gyro.R
\name{gyrodemos}
\alias{gyrodemos}
\title{Examples of the 'gyro' package}
\usage{
gyrodemos()
}
\value{
No value. The function firstly copies the demo files in a
  temporary directory. If you use RStudio, the function opens these files.
  Otherwise it prints a message giving the instructions to access to these
  files.
}
\description{
Some examples of hyperbolic polyhedra realized with the 'gyro'
  package.
}
