% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gyroreflection.R
\encoding{UTF-8}
\name{gyroreflection}
\alias{gyroreflection}
\title{Gyroreflection}
\usage{
gyroreflection(A, B, M, s, model = "U")
}
\arguments{
\item{A, B, M}{three 2D points}

\item{s}{the gyroparameter (radius of the Poincaré disk if \code{model="M"})}

\item{model}{the hyperbolic model, either \code{"M"} (Möbius model, i.e.
Poincaré model) or \code{"U"} (Ungar model, i.e. Minkowski model)}
}
\value{
A 2D point, the image of \code{M} by the reflection with respect to
  the gyroline passing through \code{A} and \code{B}.
}
\description{
Gyroreflection of a point with respect to a gyroline in
  a 2D gyrospace.
}
\examples{
library(gyro)
A <- c(1.5, 2); B <- c(2, 1)
opar <- par(mar = c(2, 2, 2, 0.5))
plot(rbind(A, B), type = "p", pch = 19, xlab = NA, ylab = NA,
     xlim = c(0.3, 2), ylim = c(0, 2.5), main = "s = 0.3")
s <- 0.3
seg <- gyrosegment(A, B, s = s, model = "U")
lines(seg, col = "blue", lwd = 2)
text(t(A), expression(italic(A)), pos = 3)
text(t(B), expression(italic(B)), pos = 3)
M <- c(1.3, 1.1)
rM <- gyroreflection(A, B, M, s = s, model = "U")
points(rbind(M, rM), type = "p", pch = 19)
text(t(M), expression(italic(M)), pos = 3)
text(t(rM), expression(italic(r(M))), pos = 3)
par(opar)
}
