% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gyro.R
\encoding{UTF-8}
\name{PhiMU}
\alias{PhiMU}
\title{Isomorphism from Ungar gyrovector space to Möbius gyrovector space}
\usage{
PhiMU(A, s = 1)
}
\arguments{
\item{A}{a point in the Ungar vector space with curvature \code{s}}

\item{s}{a positive number, the hyperbolic curvature of the Ungar
vector space}
}
\value{
The point of the Poincaré ball of radius \code{s} corresponding
  to \code{A} by isomorphism.
}
\description{
Isomorphism from the Ungar gyrovector space to
  the Möbius gyrovector space.
}
