% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gyro.R
\encoding{UTF-8}
\name{PhiUM}
\alias{PhiUM}
\title{Isomorphism from Möbius gyrovector space to Ungar gyrovector space}
\usage{
PhiUM(A, s = 1)
}
\arguments{
\item{A}{a point whose norm is lower than \code{s}}

\item{s}{positive number, the radius of the Poincaré ball}
}
\value{
The point of the Ungar gyrovector space corresponding to \code{A}
  by isomorphism.
}
\description{
Isomorphism from the Möbius gyrovector space to
  the Ungar gyrovector space.
}
