% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samtoolsIndex.R
\name{samtoolsIndex}
\alias{samtoolsIndex}
\title{Index bam files}
\usage{
samtoolsIndex(bamFiles, threads = getOption("threads", 1L))
}
\arguments{
\item{bamFiles}{A character vector of paths to bam files.}

\item{threads}{A positive integer specifying how many bams to process simultaneously.}
}
\description{
\code{samtools.index} is a simple wrapper for \code{samtools index} for indexing position-sorted bam files.
}

