% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samtoolsMerge.R
\name{samtoolsMerge}
\alias{samtoolsMerge}
\title{Merge sorted bam files}
\usage{
samtoolsMerge(bamFiles, outname, threads = getOption("threads", 1L))
}
\arguments{
\item{bamFiles}{A character vector of paths to bam files.}

\item{outname}{A vector of length one specifying the concatenated file name.}

\item{threads}{An integer specifying the number of BAM compression threads to use}
}
\description{
\code{samtools.merge} is a simple wrapper for samtools merge for merging sorted bam files.
}

