% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samtoolsSort.R
\name{samtoolsSort}
\alias{samtoolsSort}
\title{Sort bams by position or name}
\usage{
samtoolsSort(bamFiles, outputFormat = "bam", sortThreads = 1,
  threads = getOption("threads", 1L), memory = "768M", sortByName = FALSE)
}
\arguments{
\item{bamFiles}{A character vector of paths to bam files.}

\item{outputFormat}{String specifying output format: ('sam'/'bam'/'cram')}

\item{sortThreads}{A positive integer specifying the number of sorting and compression threads}

\item{threads}{A positive integer specifying how many bams to process simultaneously.}

\item{memory}{String specifying maximum memory per thread; suffix K/M/G recognized.}

\item{sortByName}{Boolean. If TRUE, reads are sorted by name. If FALSE, reads are sorted by chromosome/position}
}
\description{
\code{samtools.sort} sorts bam files either by position or read name using \code{samtools sort}.
}

