% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gyro.R
\name{gyroABt}
\alias{gyroABt}
\title{Point on a gyroline}
\usage{
gyroABt(A, B, t, s)
}
\arguments{
\item{A, B}{two distinct points}

\item{t}{a number}

\item{s}{positive number, the parameter defining the hyperbolic curvature}
}
\value{
A point.
}
\description{
Point of coordinate \code{t} on the gyroline passing through
  two given points \code{A} and \code{B}. This is \code{A} for \code{t=0}
  and this is \code{B} for \code{t=1}. For \code{t=1/2} this is the
  gyromidpoint of the gyrosegment joining \code{A} and \code{B}.
}
