% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gyro.R
\name{gyrotriangle}
\alias{gyrotriangle}
\title{Gyrotriangle in 3D space}
\usage{
gyrotriangle(A, B, C, s = 1, iterations = 5)
}
\arguments{
\item{A, B, C}{three 3D points}

\item{s}{positive number, the curvature (the smaller, the more curved)}

\item{iterations}{the gyrotriangle is constructed by iterated subdivisions,
this argument is the number of iterations}
}
\value{
A \code{\link[rgl]{mesh3d}} object.
}
\description{
3D gyrotriangle as a mesh.
}
\examples{
library(gyro)
library(rgl)
A <- c(1, 0, 0); B <- c(0, 1, 0); C <- c(0, 0, 1)
ABC <- gyrotriangle(A, B, C, s = 0.3)
view3d(30, 30, zoom = 0.75)
shade3d(ABC, color = "navy", specular = "cyan")
}
