% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gyro.R
\name{gyrotube}
\alias{gyrotube}
\title{Gyrotube (tubular gyrosegment)}
\usage{
gyrotube(A, B, s = 1, n = 100, radius, sides = 90, caps = FALSE)
}
\arguments{
\item{A, B}{distinct 3D points}

\item{s}{positive number, the curvature (higher value, less curved)}

\item{n}{number of points forming the gyrosegment}

\item{radius}{radius of the tube around the gyrosegment}

\item{sides}{number of sides in the polygon cross section}

\item{caps}{Boolean, whether to put caps on the ends of the tube}
}
\value{
A \code{\link[rgl]{mesh3d}} object.
}
\description{
Tubular gyrosegment joining two given 3D points.
}
\examples{
library(gyro)
library(rgl)
A <- c(1, 2, 0); B <- c(1, 1, 0)
tube <- gyrotube(A, B, s = 0.2, radius = 0.02)
shade3d(tube, color = "orangered")
}
