% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdelaunay.R
\encoding{UTF-8}
\name{hdelaunay}
\alias{hdelaunay}
\title{Hyperbolic Delaunay triangulation}
\usage{
hdelaunay(points, isolations = FALSE, exact = FALSE)
}
\arguments{
\item{points}{points in the unit disk given as a numeric matrix with
two columns}

\item{isolations}{Boolean, whether to identify isolated vertices and edges}

\item{exact}{Boolean, whether to perform exact calculations; slower but
more accurate}
}
\value{
A list with four fields \code{vertices}, \code{edges},
  \code{triangles} and \code{ntriangles}, and two additional fields
  \code{mvertices} and \code{medges} if \code{isolations=TRUE}, giving
  the non-isolated vertices and edges ("m" for "multivalent"). The input
  \code{points} matrix and the output \code{vertices} matrix are the same
  up to the order of the rows.
}
\description{
Computes the hyperbolic Delaunay triangulation of a set of
  points in the Poincaré disk.
}
\examples{
library(gyro)
library(uniformly)
set.seed(666)
points <- runif_in_sphere(10L, d = 2)
hdelaunay(points)
}
\seealso{
\code{\link{plotHdelaunay}}
}
