% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdelaunay.R
\name{plotHdelaunay}
\alias{plotHdelaunay}
\title{Plot hyperbolic Delaunay triangulation}
\usage{
plotHdelaunay(
  hdel,
  remove = NULL,
  vertices = TRUE,
  edges = TRUE,
  circle = TRUE,
  color = "distinct",
  hue = "random",
  luminosity = "random"
)
}
\arguments{
\item{hdel}{an output of \code{\link{hdelaunay}}}

\item{remove}{what you want to remove, \code{NULL} for nothing,
\code{"ivertices"} for the isolated vertices, \code{"iedges"}
for the isolated edges or \code{c("ivertices", "iedges")} for
the isolated vertices and the isolated edges; if not \code{NULL},
this assumes you ran \code{\link{hdelaunay}} with the
option \code{isolations=TRUE}}

\item{vertices}{Boolean, whether to plot the vertices}

\item{edges}{Boolean, whether to plot the edges}

\item{circle}{Boolean, whether to plot the unit circle}

\item{color}{this argument controls the colors of the triangles; it can be
\code{NA} for no color, \code{"random"} for random colors generated
with \code{\link[randomcoloR]{randomColor}}, \code{"distinct"} for
distinct colors generated with
\code{\link[randomcoloR]{distinctColorPalette}}, a single color or
a vector of colors}

\item{hue, luminosity}{passed to \code{\link[randomcoloR]{randomColor}}
if \code{color="random"}}
}
\value{
No returned value.
}
\description{
Plot a hyperbolic Delaunay triangulation obtained
  with \code{\link{hdelaunay}}.
}
\examples{
library(gyro)
library(uniformly)
set.seed(666)
points <- runif_in_sphere(35L, d = 2)
hdel <- hdelaunay(points)
plotHdelaunay(hdel)
}
