% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdelaunay.R
\name{plotHdelaunay}
\alias{plotHdelaunay}
\title{Plot hyperbolic Delaunay triangulation}
\usage{
plotHdelaunay(
  hdel,
  remove = NULL,
  vertices = TRUE,
  edges = TRUE,
  circle = TRUE,
  color = "random",
  hue = "random",
  luminosity = "random"
)
}
\arguments{
\item{hdel}{an output of \code{\link{hdelaunay}}}

\item{remove}{\code{NULL}, \code{"ivertices"}, \code{"iedges"} or \code{c("ivertices", "iedges")}}

\item{vertices}{Boolean}

\item{edges}{Boolean}

\item{circle}{Boolean,}

\item{color}{either}

\item{hue, luminosity}{passed to \code{\link[randomcoloR]{randomColor}}}
}
\value{
No returned value.
}
\description{
Plot a hyperbolic Delaunay triangulation obtained
  with \code{\link{hdelaunay}}.
}
\examples{
library(gyro)
library(uniformly)
set.seed(666)
points <- runif_in_sphere(25L, d = 2)
hdel <- hdelaunay(points)
plotHdelaunay(hdel)
}
